RF: minor of DialogUI to allow for non-interactive ConsoleLog
